+++ packages/react/src/components/index.ts
@@ -36,6 +36,7 @@ export * from "./MannerTemp";
 export * from "./MannerTempBadge";
 export * from "./MenuSheet";
 export * from "./NotificationBadge";
+export * from "./PageBanner";
 export * from "./Portal";
 export * from "./ProgressCircle";
 export * from "./PullToRefresh";